backend/deploy/requirements.txt
aiohappyeyeballs==2.4.4 ; python_version >= "3.12" and python_version < "4.0"
aiohttp==3.11.9 ; python_version >= "3.12" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
attrs==24.2.0 ; python_version >= "3.12" and python_version < "4.0"
blinker==1.8.2 ; python_version >= "3.12" and python_version < "4.0"
cachecontrol==0.14.0 ; python_version >= "3.12" and python_version < "4.0"
cachetools==5.5.0 ; python_version >= "3.12" and python_version < "4.0"
certifi==2024.8.30 ; python_version >= "3.12" and python_version < "4.0"
cffi==1.17.1 ; python_version >= "3.12" and python_version < "4.0" and platform_python_implementation != "PyPy"
charset-normalizer==3.3.2 ; python_version >= "3.12" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.12" and python_version < "4.0"
cloudevents==1.11.0 ; python_version >= "3.12" and python_version < "4"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and (sys_platform == "win32" or platform_system == "Windows")
colorlog==6.8.2 ; python_version >= "3.12" and python_version < "4.0"
cryptography==43.0.1 ; python_version >= "3.12" and python_version < "4.0"
deprecation==2.1.0 ; python_version >= "3.12" and python_version < "4"
docstring-parser==0.16 ; python_version >= "3.12" and python_version < "4.0"
firebase-admin==6.5.0 ; python_version >= "3.12" and python_version < "4.0"
flask-cors==5.0.0 ; python_version >= "3.12" and python_version < "4.0"
flask==3.0.3 ; python_version >= "3.12" and python_version < "4.0"
frozenlist==1.5.0 ; python_version >= "3.12" and python_version < "4.0"
functions-framework==3.8.2 ; python_version >= "3.12" and python_version < "4"
google-ai-generativelanguage==0.6.10 ; python_version >= "3.12" and python_version < "4.0"
google-api-core==2.21.0 ; python_version >= "3.12" and python_version < "4.0"
google-api-core[grpc]==2.21.0 ; python_version >= "3.12" and python_version < "4.0"
google-api-python-client==2.149.0 ; python_version >= "3.12" and python_version < "4.0"
google-auth-httplib2==0.2.0 ; python_version >= "3.12" and python_version < "4.0"
google-auth==2.35.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-aiplatform==1.70.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-bigquery==3.26.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-core==2.4.1 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-firestore==2.19.0 ; python_version >= "3.12" and python_version < "4.0" and platform_python_implementation != "PyPy"
google-cloud-resource-manager==1.13.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-storage==2.18.2 ; python_version >= "3.12" and python_version < "4.0"
google-crc32c==1.6.0 ; python_version >= "3.12" and python_version < "4.0"
google-generativeai==0.8.3 ; python_version >= "3.12" and python_version < "4.0"
google-resumable-media==2.7.2 ; python_version >= "3.12" and python_version < "4.0"
googleapis-common-protos==1.65.0 ; python_version >= "3.12" and python_version < "4.0"
googleapis-common-protos[grpc]==1.65.0 ; python_version >= "3.12" and python_version < "4.0"
grpc-google-iam-v1==0.13.1 ; python_version >= "3.12" and python_version < "4.0"
grpcio-status==1.66.2 ; python_version >= "3.12" and python_version < "4.0"
grpcio==1.66.2 ; python_version >= "3.12" and python_version < "4.0"
gunicorn==23.0.0 ; python_version >= "3.12" and python_version < "4" and platform_system != "Windows"
httplib2==0.22.0 ; python_version >= "3.12" and python_version < "4.0"
idna==3.10 ; python_version >= "3.12" and python_version < "4.0"
iniconfig==2.0.0 ; python_version >= "3.12" and python_version < "4.0"
itsdangerous==2.2.0 ; python_version >= "3.12" and python_version < "4.0"
jinja2==3.1.4 ; python_version >= "3.12" and python_version < "4.0"
joblib==1.4.2 ; python_version >= "3.12" and python_version < "4.0"
markupsafe==2.1.5 ; python_version >= "3.12" and python_version < "4.0"
msgpack==1.1.0 ; python_version >= "3.12" and python_version < "4.0"
multidict==6.1.0 ; python_version >= "3.12" and python_version < "4.0"
numpy==2.1.2 ; python_version >= "3.12" and python_version < "4.0"
openai==0.28.0 ; python_version >= "3.12" and python_version < "4.0"
packaging==24.1 ; python_version >= "3.12" and python_version < "4.0"
pandas==2.2.3 ; python_version >= "3.12" and python_version < "4.0"
pdfminer-six==20231228 ; python_version >= "3.12" and python_version < "4.0"
pdfplumber==0.11.4 ; python_version >= "3.12" and python_version < "4.0"
pdfrw==0.4 ; python_version >= "3.12" and python_version < "4.0"
pillow==10.4.0 ; python_version >= "3.12" and python_version < "4.0"
pluggy==1.5.0 ; python_version >= "3.12" and python_version < "4.0"
propcache==0.2.1 ; python_version >= "3.12" and python_version < "4.0"
proto-plus==1.25.0 ; python_version >= "3.12" and python_version < "4.0"
protobuf==5.28.2 ; python_version >= "3.12" and python_version < "4.0"
pyasn1-modules==0.4.1 ; python_version >= "3.12" and python_version < "4.0"
pyasn1==0.6.1 ; python_version >= "3.12" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.12" and python_version < "4.0" and platform_python_implementation != "PyPy"
pycryptodome==3.21.0 ; python_version >= "3.12" and python_version < "4.0"
pydantic-core==2.23.4 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.9.2 ; python_version >= "3.12" and python_version < "4.0"
pyjwt[crypto]==2.9.0 ; python_version >= "3.12" and python_version < "4.0"
pymupdf==1.24.11 ; python_version >= "3.12" and python_version < "4.0"
pyparsing==3.1.4 ; python_version >= "3.12" and python_version < "4.0"
pypdf==5.1.0 ; python_version >= "3.12" and python_version < "4.0"
pypdfium2==4.30.0 ; python_version >= "3.12" and python_version < "4.0"
pytest==8.3.3 ; python_version >= "3.12" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.12" and python_version < "4.0"
pytz==2024.2 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.12" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.12" and python_version < "4"
scikit-learn==1.5.2 ; python_version >= "3.12" and python_version < "4.0"
scipy==1.14.1 ; python_version >= "3.12" and python_version < "4.0"
shapely==2.0.6 ; python_version >= "3.12" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.12" and python_version < "4.0"
threadpoolctl==3.5.0 ; python_version >= "3.12" and python_version < "4.0"
tqdm==4.66.6 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.12" and python_version < "4.0"
tzdata==2024.2 ; python_version >= "3.12" and python_version < "4.0"
uritemplate==4.1.1 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.2.3 ; python_version >= "3.12" and python_version < "4.0"
watchdog==6.0.0 ; python_version >= "3.12" and python_version < "4"
werkzeug==3.0.4 ; python_version >= "3.12" and python_version < "4.0"
yarl==1.18.3 ; python_version >= "3.12" and python_version < "4.0"
